Bump Acontplus.Core to version 1.2.9 and update XML documentation for AuditableEntity. Refactor project structure by introducing new entities and extensions for improved error handling and response management. Enhance API response helpers for better success and error messaging.
