+++ src/Acontplus.Core/Acontplus.Core.csproj
@@ -5,7 +5,7 @@
     <Nullable>enable</Nullable>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageId>Acontplus.Core</PackageId>
-    <Version>1.2.8</Version>
+    <Version>1.2.9</Version>
     <Authors>Ivan Paz</Authors>
     <Company>Acontplus S.A.S.</Company>
     <Description>Modern .NET 9+ foundational components for enterprise applications. Includes domain-driven design patterns, comprehensive error handling, specification pattern, and modern C# features.</Description>
