+++ src/Acontplus.Core/Api/Helpers/ApiResponseHelpers.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net;
 
-namespace Acontplus.Core.Helpers;
+namespace Acontplus.Core.Api.Helpers;
 
 public static class ApiResponseHelpers
 {
