+++ src/Acontplus.Core/Domain/Enums/DeviceType.cs
@@ -1,4 +1,4 @@
-﻿namespace Acontplus.Core.Enums;
+﻿namespace Acontplus.Core.Domain.Enums;
 
 public enum DeviceType
 {
