+++ src/Acontplus.Core/Domain/Enums/ResponseStatus.cs
@@ -1,4 +1,4 @@
-﻿namespace Acontplus.Core.Enums;
+﻿namespace Acontplus.Core.Domain.Enums;
 
 public enum ResponseStatus
 {
