+++ src/Acontplus.Core/Domain/Extensions/DomainWarningsExtensions.cs
@@ -1,4 +1,6 @@
-﻿namespace Acontplus.Core.Extensions;
+﻿using Acontplus.Core.Domain.Common.Results;
+
+namespace Acontplus.Core.Domain.Extensions;
 
 public static class DomainWarningsExtensions
 {
