+++ src/Acontplus.Core/Domain/Extensions/PagedResultExtensions.cs
@@ -1,4 +1,4 @@
-﻿namespace Acontplus.Core.Extensions;
+﻿namespace Acontplus.Core.Domain.Extensions;
 
 
 public static class PagedResultExtensions
